requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python-headless==4.4.0.46
Pillow==8.0.1
protobuf==3.14.0
tensorflow-cpu==2.4.0
Werkzeug==1.0.1
